Python 3 support, bug fixes and pep8
This commit changes pretty much all files in this package.

First, a number of bugs with crawler object are fixed. The code is made
compatible with python 3, and steps were taken to make more readable,
less verbose and more maintainable.

Names, specially captalization are changed across the board to comply
with python's style.

Companies.txt is no more, instead the data was transfered to a python
file, and loaded in a dictionary as named tuple. This makes accessing
them from within a python program considerable easier and more natural.

The test.py script was also completely rewritten, being now much shorter
and straight to the point, basically only showcasing the basic
functionality of the crawler.

The version was updated to 0.2.0, since the changes are big enough to
justify it.