Pass user agent to CIK lookup (#297)
* Pass user agent to cik lookup

The CIK lookup (central index key) uses the EDGAR interface to get
tickers and company names. The lookup is separate from the core REST
interface which is where most of the EDGAR calls are made. The CIK
lookup also requires a valid user agent being subject to the same filter
rules as the others.

SEC blocks requests where user agent contains the string
"python-request". There may be other filters present but not documented.

* fixup! Pass user agent to cik lookup